# smd-vital-backend/docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: SMD Vital - Microservices API
  description: |
    API completa para la plataforma de servicios médicos domiciliarios SMD Vital Bogotá.
    
    ## Servicios Disponibles
    
    - **Authentication Service**: Gestión de autenticación y autorización
    - **User Service**: Gestión de usuarios y perfiles
    - **Appointment Service**: Programación y gestión de citas médicas
    - **Notification Service**: Envío de notificaciones multicanal
    - **Medical Records Service**: Historiales médicos y documentos
    - **Payment Service**: Procesamiento de pagos y facturación
    
    ## Autenticación
    
    Todos los endpoints (excepto públicos) requieren autenticación JWT.
    Incluye el token en el header: `Authorization: Bearer <token>`
    
    ## Códigos de Estado
    
    - `200` - Éxito
    - `201` - Creado exitosamente
    - `400` - Error en la solicitud
    - `401` - No autorizado
    - `403` - Prohibido
    - `404` - No encontrado
    - `422` - Error de validación
    - `500` - Error interno del servidor
  version: 1.0.0
  contact:
    name: SMD Vital Backend Team
    email: dev@smdvitalbogota.com
    url: https://smdvitalbogota.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo Local (API Gateway)
  - url: http://localhost:8001
    description: Auth Service Directo
  - url: http://localhost:8002
    description: User Service Directo
...
paths:
...
  /appointments:
    get:
      tags:
        - Appointments
      summary: Listar citas
      description: Obtener lista de citas del usuario
      operationId: listAppointments
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por estado de cita
          schema:
            type: string
            enum: [pending, confirmed, in_progress, completed, cancelled]
        - name: page
          in: query
          description: Número de página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Elementos por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Lista de citas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'

    post:
      tags:
        - Appointments
      summary: Crear nueva cita
      description: Programar una nueva cita médica
      operationId: createAppointment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreate'
      responses:
        '201':
          description: Cita creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
  /appointments/{appointmentId}:
    get:
      tags:
        - Appointments
      summary: Obtener cita específica
      description: Obtener detalles de una cita específica
      operationId: getAppointment
      security:
        - BearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          description: ID de la cita
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalles de la cita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          description: Cita no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Appointments
      summary: Actualizar cita
      description: Actualizar información de una cita existente
      operationId: updateAppointment
      security:
        - BearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          description: ID de la cita
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdate'
      responses:
        '200':
          description: Cita actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'

    delete:
      tags:
        - Appointments
      summary: Cancelar cita
      description: Cancelar una cita programada
      operationId: cancelAppointment
      security:
        - BearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          description: ID de la cita
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cita cancelada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Medical Records Service Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido del endpoint de login
  schemas:
...
    AppointmentCreate:
      type: object
      required:
        - service_type
        - scheduled_date
        - address
      properties:
        service_type:
          type: string
          enum: [consulta_general, inyectologia, toma_muestras, curaciones, signos_vitales]
          description: Tipo de servicio médico
        scheduled_date:
          type: string
          format: date-time
          description: Fecha y hora programada
        address:
          type: string
          description: Dirección donde se realizará el servicio
        notes:
          type: string
          description: Notas adicionales
        emergency:
          type: boolean
          default: false
          description: Es una emergencia
    AppointmentUpdate:
      type: object
      properties:
        scheduled_date:
          type: string
          format: date-time
        address:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID de la cita
        patient_id:
          type: string
          format: uuid
          description: ID del paciente
        professional_id:
          type: string
          format: uuid
          description: ID del profesional asignado
        service_type:
          type: string
          description: Tipo de servicio
        scheduled_date:
          type: string
          format: date-time
          description: Fecha programada
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        address:
          type: string
          description: Dirección del servicio
        notes:
          type: string
          description: Notas
        emergency:
          type: boolean
          description: Es emergencia
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AppointmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Página actual
        limit:
          type: integer
          description: Elementos por página
        total:
          type: integer
          description: Total de elementos
        pages:
          type: integer
          description: Total de páginas

    # Response Schemas
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
          description: Mensaje de éxito
        data:
          type: object
          description: Datos adicionales
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
          description: Mensaje de error
        error_code:
          type: string
          description: Código de error
        details:
          type: object
          description: Detalles adicionales del error
...